TAR:12633

+++ api.mstiADM.SDK.csharp/ViewModels/ClientPackVM/ConfigDFeVMPack.cs
@@ -11,21 +11,17 @@ namespace api.mstiADM.SDK.csharp.ViewModels.ClientPackVM
         [Key(0)]
         public string SerieAtual { get; set; }
 
-
         /// <summary> 
         /// <para>Próximo número a ser enviado</para> 
         /// </summary> 
         [Key(1)]
-        public long ProximoLote { get; set; }
-
+        public int ProximoLote { get; set; }
 
         /// <summary> 
         /// <para>Próximo lote a ser enviado</para> 
         /// </summary> 
         [Key(2)]
-        public long ProximoNumero { get; set; }
-
-
+        public int ProximoNumero { get; set; }
 
         public ConfigDFeVM ToConfigDFeVM()
         {
